move getTwins and getSiblings to utils
